Move Mark Attendance package to Meetup module
