+++ app/src/main/java/dev/android/peru/modules/meetup/markAttendance/MarkAttendanceFragment.kt
@@ -1,4 +1,4 @@
-package dev.android.peru.modules.search.markAttendance
+package dev.android.peru.modules.meetup.markAttendance
 
 import android.app.AlertDialog
 import android.os.Bundle
