+++ app/src/main/java/dev/android/peru/modules/meetup/markAttendance/MarkAttendanceViewModel.kt
@@ -1,4 +1,4 @@
-package dev.android.peru.modules.search.markAttendance
+package dev.android.peru.modules.meetup.markAttendance
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
